requirements.txt
pydantic~=2.1.1
psycopg2-binary==2.9.7
requests~=2.31.0
pendulum>=2.0
apache-airflow-providers-docker
